make pagination hooks store filter & query params in URL, not just pagination params
Make `usePageSwitcherPagination` and `useShowMorePaginationUrl` support storing filter and query params, not just pagination params, in the URL. This is commonly desired behavior, and there are many ways we do it across the codebase. This attempts to standardize how it's done. It does not update all places this is done to standardize them yet.